% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvoa_weekly.R
\name{dvoa_weekly}
\alias{dvoa_weekly}
\title{scrapes the football-outsiders weekly dvoa pages to pull their model's assessment of the previous week's performance
the unfortunate fact is that the table required table is not in the same position every week, so you need to visit the page, find the table you want and denote it's number under table number
the only other piece you need to provide is the week_number}
\usage{
dvoa_weekly(week_num, table_num)
}
\arguments{
\item{week_num}{number week of NFL play you want to know dvoa of}

\item{table_num}{the ordinal value of where the table you want lives on the page relative to the other tables}
}
\value{
a tibble denoting the dvoa for the week designated
}
\description{
scrapes the football-outsiders weekly dvoa pages to pull their model's assessment of the previous week's performance
the unfortunate fact is that the table required table is not in the same position every week, so you need to visit the page, find the table you want and denote it's number under table number
the only other piece you need to provide is the week_number
}
