% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridiron.R
\name{gridiron}
\alias{gridiron}
\title{After using oddsmakr:odds to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. share for each team by game using this function}
\usage{
gridiron(d)
}
\arguments{
\item{d}{dataframe or tibble you want to pass to the function}
}
\value{
a patchwork plot showing all spread and spread_share for each game
}
\description{
After using oddsmakr:odds to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. share for each team by game using this function
}
