% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridiron_ou_all.R
\name{gridiron_ou}
\alias{gridiron_ou}
\title{After using oddsmakr::odds_consensus to scrape data and you've compiled it into a dataframe, you can plot a time-series of over-under vs. % share of bet volume for each team by game using this function}
\usage{
gridiron_ou(dataset, week_no)
}
\arguments{
\item{dataset}{dataframe or tibble containing public consensus share of bets and over-under for multiples games for a given week pulled more than once by odds_consensus function from oddsmaker package}

\item{week_no}{the week number to visualize}
}
\value{
a patchwork plot showing all spread and spread_share for each game
}
\description{
After using oddsmakr::odds_consensus to scrape data and you've compiled it into a dataframe, you can plot a time-series of over-under vs. % share of bet volume for each team by game using this function
}
