% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_season_schedule.R
\name{nfl_season_schedule}
\alias{nfl_season_schedule}
\title{scrapes season schedule from pro-football-reference.com}
\usage{
nfl_season_schedule(year)
}
\arguments{
\item{year}{the season you'd like to pull}
}
\value{
a tibble containing the full or a portion of the NFL schedule for a particular season
}
\description{
scrapes season schedule from pro-football-reference.com
}
