% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_consensus.R
\name{odds_consensus}
\alias{odds_consensus}
\title{scrapes the 'odds consensus' page from oddsshark.com to provide insight on where the public is betting on any given game
all you need to do is enter the amount of games for that week listed on the oddsshark consensus page (https://www.oddsshark.com/nfl/consensus-picks) and it will generate a table that lists the different game, their respective spreads at that point-in-time, and what share of wagers have gone in either direction}
\usage{
odds_consensus(week_num, num_games)
}
\arguments{
\item{week_num}{number week of NFL play are we currently in}

\item{num_games}{number of games listed on oddsshark.com consensus site that have percent shares listed}
}
\value{
a tibble of all games for that week
}
\description{
scrapes the 'odds consensus' page from oddsshark.com to provide insight on where the public is betting on any given game
all you need to do is enter the amount of games for that week listed on the oddsshark consensus page (https://www.oddsshark.com/nfl/consensus-picks) and it will generate a table that lists the different game, their respective spreads at that point-in-time, and what share of wagers have gone in either direction
}
