% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scores.R
\name{pfr_scores}
\alias{pfr_scores}
\title{scrapes the final scores page on https://www.pro-football-reference.com to get each week's final scores
list the week #, the # of games, and the URL to gather final scores for each game}
\usage{
pfr_scores(week_num, num_games, game_url)
}
\arguments{
\item{week_num}{number week of NFL play are we currently in}

\item{num_games}{number of games listed on oddsshark.com consensus site that have percent shares listed}

\item{game_url}{provide the URL for the week of games on profootball reference}
}
\value{
a tibble of all games for that week
}
\description{
scrapes the final scores page on https://www.pro-football-reference.com to get each week's final scores
list the week #, the # of games, and the URL to gather final scores for each game
}
