% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_share_plot.R
\name{pick_share_plot}
\alias{pick_share_plot}
\title{After using oddsmaker::picks_consensus to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. share for each team by game using this function}
\usage{
pick_share_plot(dat)
}
\arguments{
\item{dat}{dataframe or tibble containing public consensus share of bets and game line for a single game for a given week pulled more than once by odds_consensus function from oddsmaker package}
}
\value{
a times-series chart showing public consensus and lines movement
}
\description{
After using oddsmaker::picks_consensus to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. share for each team by game using this function
}
