% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridiron_all.R
\name{gridiron_all}
\alias{gridiron_all}
\title{After using oddsmakr:odds to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. % share of bet volume for each team by game using this function}
\usage{
gridiron_all(dat)
}
\arguments{
\item{dat}{dataframe or tibble containing public consensus share of bets and game line for multiples games for a given week pulled more than once by odds_consensus function from oddsmaker package}
}
\value{
a patchwork plot showing all spread and spread_share for each game
}
\description{
After using oddsmakr:odds to scrape data and you've compiled it into a dataframe, you can plot a time-series of spreads vs. % share of bet volume for each team by game using this function
}
