% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picks_consensus.R
\name{picks_consensus}
\alias{picks_consensus}
\title{updated version of odds_consensus, scraping oddsmaker 'public conensus picks' site}
\usage{
picks_consensus()
}
\arguments{
\item{NULL}{}
}
\value{
a tibble of picks consensus across all games on https://www.oddsshark.com/nfl/consensus-picks at a moment in time
}
\description{
updated version of odds_consensus, scraping oddsmaker 'public conensus picks' site
}
